% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcdc-package.R
\docType{package}
\name{bcdc-package}
\alias{bcdc}
\alias{bcdc-package}
\title{bcdc: Bayesian community detection for networks with covariates}
\description{
An implementation of the a Bayesian community detection model for networks with covariates.
}
\author{
\strong{Maintainer}: Arash A. Amini \email{aaamini@ucla.edu} (\href{https://orcid.org/0000-0002-2808-8310}{ORCID})

Authors:
\itemize{
  \item Luyi Shen
  \item Nathaniel Josephs
  \item Lizhen Lin
}

}
